+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Direct/indirect imports and listing imported packages
+- pin/unpin packages and unimport --all
 
 ## [0.1.4] - 2025-11-24